// src/app/log-action-type.enum.ts
export enum LogActionType{
    LOGIN = 'LOGIN',
    SIGNUP = 'SIGNUP',
    PASSWORD_RESET = 'FORGOT_PASSWORD',
    UNENROLL_FAILED = 'UNENROLL_FAILED',
    ASSIGN_GRADE_FAILED = 'ASSIGN_GRADE_FAILED',
    LOGOUT = 'LOGOUT',
    ENROLL = 'ENROLL',
    UNENROLL = 'UNENROLL',
    ASSIGN_GRADE = 'ASSIGN_GRADE',
    CREATE_COURSE = 'CREATE_COURSE',
    UPDATE_COURSE = 'UPDATE_COURSE',
    DELETE_COURSE = 'DELETE_COURSE',
    DELETE_USER = 'DELETE_USER',
    ASSIGN_ROLE = 'ASSIGN_ROLE',
    UPDATE_COURSE_FAILED = 'UPDATE_COURSE_FAILED',
    DELETE_COURSE_FAILED = 'DELETE_COURSE_FAILED',
    GENERIC = 'GENERIC'
  }